refactor RecipeCard comp
